Trying to get e6v2/oe:default to compile
