--- keyboards/e6v2/keymaps/amnesia0287/config.h
@@ -1,1 +1,0 @@
-#include "../../config.h"
