--- keyboards/e6v2/keymaps/default/config.h
@@ -1,1 +1,0 @@
-#include "../../config.h"
