+++ keyboards/e6v2/le/le.h
@@ -1,7 +1,7 @@
 #ifndef LE_H
 #define LE_H
 
-#include "quantum.h"
+#include "../e6v2.h"
 
 #define LAYOUT( \
     K00, K01, K02, K03, K04, K05, K06, K07, K08, K09, K0A, K0B, K0C, K0D, K0E, \
