+++ keyboards/e6v2/oe/keymaps/amnesia0287/readme.md
